A4/backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-crispy-forms==2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2022.7.1
reportlab==3.6.12
sqlparse==0.4.3
tomli==2.0.1
